services/real-time-events-collector/requirements.txt
Flask==2.0.1
kafka-python==2.0.2
Werkzeug==2.0.3
flask-cors==3.0.10